common CommandAction
