+++ src/main/java/me/hsgamer/bettergui/action/type/OpAction.java
@@ -1,13 +1,7 @@
 package me.hsgamer.bettergui.action.type;
 
-import me.hsgamer.bettergui.BetterGUI;
-import me.hsgamer.bettergui.api.action.CommandAction;
 import me.hsgamer.bettergui.builder.ActionBuilder;
-import me.hsgamer.hscore.task.element.TaskProcess;
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
-
-import java.util.UUID;
 
 public class OpAction extends CommandAction {
   public OpAction(ActionBuilder.Input input) {
@@ -15,25 +9,16 @@ public class OpAction extends CommandAction {
   }
 
   @Override
-  public void accept(UUID uuid, TaskProcess process) {
-    String replacedString = getFinalCommand(uuid);
-    Player player = Bukkit.getPlayer(uuid);
-    if (player == null) {
-      process.next();
-      return;
-    }
-    Bukkit.getScheduler().runTask(BetterGUI.getInstance(), () -> {
-      if (player.isOp()) {
-        player.chat(replacedString);
-      } else {
-        try {
-          player.setOp(true);
-          player.chat(replacedString);
-        } finally {
-          player.setOp(false);
-        }
+  protected void accept(Player player, String command) {
+    if (player.isOp()) {
+      player.chat(command);
+    } else {
+      try {
+        player.setOp(true);
+        player.chat(command);
+      } finally {
+        player.setOp(false);
       }
-      process.next();
-    });
+    }
   }
 }
